stop reading the body once it has been closed
We still looped through all entries in a body until that streamer
finished all data.
Once a body has been closed, it means it's connection isn't open
anymore, and we can stop reading.

Among a lot of disconnections, this could have caused latencies, as
there were a lot of bodies open, all of which acquiring a lock on the
mutex to know if they could close themselves or not.